+++ typedoc.config.mjs
@@ -74,7 +74,6 @@ const typedocPluginReplaceTextOptions = {
 /** @type {Partial<import("typedoc").TypeDocOptions>} */
 const config = {
   out: './.typedoc/docs',
-  json: './.typedoc/docs.json',
   entryPointStrategy: 'packages',
   plugin: [
     'typedoc-plugin-replace-text',