+++ packages/expo-passkeys/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @clerk/expo-passkeys
 
+## 0.1.27
+
+### Patch Changes
+
+- Updated dependencies [[`466ed13`](https://github.com/clerk/javascript/commit/466ed136af73b59b267d92ad3296039d1c3a4fcc)]:
+  - @clerk/types@4.49.2
+  - @clerk/shared@3.2.1
+
 ## 0.1.26
 
 ### Patch Changes